[Added] - stats of breaked and placed per Block type

+++ src/org/rage/pluginstats/stats/Mob.java
@@ -45,7 +45,7 @@ public class Mob {
 		this.mName = mName;
 	}
 	
-	public void setNumMobKilled(int mNumKilled) {
+	public void setNumMobKilled(long mNumKilled) {
 		this.mNumKilled = mNumKilled;
 	}
